requirements.txt
fastapi
uvicorn
openai==0.28
pinecone
sentence-transformers
langchain
python-dotenv
